small fixes based on review
